Move txout from transaction to seperate file.
